# app/routers/marketing_fixed.py
from fastapi import APIRouter, Query
from typing import Optional

router = APIRouter(prefix="/api/marketing", tags=["MARKETING"])

# MARKETING - campanhas, promoções e comunicação - Compatível com Meep

# Constantes
OWNER_ID_DESC = "ID do proprietário"
NAME_FILTER_DESC = "Nome para filtro"

# ============================================================================
# FIDELIDADE (Fidelity API) - Baseado em fidelity-api.meep.cloud
# ============================================================================

@router.get("/fidelidade/clientes/historico")
async def buscar_historico_mudancas_clientes(
    owner_id: str = Query(..., description=OWNER_ID_DESC),
    change_history_type: int = Query(1, description="Tipo de histórico de mudança")
):
    """Busca histórico de mudanças de clientes fidelizados - Compatível com Meep"""
    return {
        "status": "success",
        "module": "fidelidade",
        "endpoint": "historico_clientes",
        "owner_id": owner_id,
        "change_history_type": change_history_type,
        "data": []
    }

@router.get("/fidelidade/hierarquia/{owner_id}")
async def buscar_hierarquia_fidelidade(owner_id: str):
    """Busca hierarquia do programa de fidelidade - Compatível com Meep"""
    return {
        "status": "success",
        "module": "fidelidade", 
        "endpoint": "hierarquia",
        "owner_id": owner_id,
        "hierarchy": [
            {
                "level": 1,
                "name": "PRATA BROZE",
                "min_points": 0,
                "customers_count": 8064,
                "benefits": []
            }
        ]
    }

@router.get("/fidelidade/troca-moedas/{local_id}")
async def buscar_troca_moedas_local(local_id: str):
    """Busca configuração de troca de moedas por local - Compatível com Meep"""
    return {
        "status": "success",
        "module": "fidelidade",
        "endpoint": "troca_moedas",
        "local_id": local_id,
        "exchange_config": {
            "rate": 1.0,
            "min_amount": 10,
            "max_amount": 1000
        }
    }

@router.get("/fidelidade/pontos/{customer_id}")
async def buscar_pontos_cliente(customer_id: str):
    """Busca pontos acumulados do cliente - Compatível com Meep"""
    return {
        "status": "success",
        "module": "fidelidade",
        "endpoint": "pontos_cliente",
        "customer_id": customer_id,
        "points": {
            "current": 2500,
            "total_earned": 5000,
            "total_used": 2500,
            "expiring_soon": 100
        }
    }

# ============================================================================
# CRM (CRM API) - Baseado em crm-api.meep.cloud
# ============================================================================

@router.get("/crm/customers")
async def listar_clientes_crm(
    owner_id: str = Query(..., description=OWNER_ID_DESC),
    page: int = Query(1, description="Página"),
    limit: int = Query(10, description="Limite por página")
):
    """Lista clientes do CRM - Compatível com Meep"""
    return {
        "status": "success",
        "module": "crm",
        "endpoint": "customers",
        "owner_id": owner_id,
        "pagination": {"page": page, "limit": limit, "total": 8064},
        "customers": []
    }

@router.get("/crm/campaigns")
async def listar_campanhas_crm(
    owner_id: str = Query(..., description=OWNER_ID_DESC),
    status: Optional[str] = Query(None, description="Status da campanha")
):
    """Lista campanhas de marketing do CRM - Compatível com Meep"""
    return {
        "status": "success",
        "module": "crm",
        "endpoint": "campaigns",
        "owner_id": owner_id,
        "filter_status": status,
        "campaigns": []
    }

@router.post("/crm/campaigns")
async def criar_campanha_crm(
    owner_id: str = Query(..., description=OWNER_ID_DESC)
):
    """Cria nova campanha de marketing - Compatível com Meep"""
    return {
        "status": "success",
        "module": "crm",
        "endpoint": "create_campaign",
        "owner_id": owner_id,
        "campaign_id": "camp_123456"
    }

@router.get("/crm/events")
async def listar_eventos_crm(
    owner_id: str = Query(..., description=OWNER_ID_DESC)
):
    """Lista eventos para campanhas de CRM - Compatível com Meep"""
    return {
        "status": "success",
        "module": "crm",
        "endpoint": "events",
        "owner_id": owner_id,
        "events": []
    }

@router.get("/crm/event-categories")
async def listar_categorias_eventos():
    """Lista categorias de eventos - Compatível com Meep"""
    return {
        "status": "success",
        "module": "crm",
        "endpoint": "event_categories",
        "categories": [
            {"id": 1, "name": "Show", "color": "#FF5722"},
            {"id": 2, "name": "Festival", "color": "#2196F3"},
            {"id": 3, "name": "Teatro", "color": "#9C27B0"}
        ]
    }

@router.get("/crm/ticket-categories")
async def listar_categorias_ingressos():
    """Lista categorias de ingressos - Compatível com Meep"""
    return {
        "status": "success",
        "module": "crm",
        "endpoint": "ticket_categories",
        "categories": [
            {"id": 1, "name": "VIP", "price": 150.00},
            {"id": 2, "name": "Pista", "price": 80.00},
            {"id": 3, "name": "Camarote", "price": 200.00}
        ]
    }

@router.get("/crm/saved-filters")
async def listar_filtros_salvos(
    owner_id: str = Query(..., description=OWNER_ID_DESC)
):
    """Lista filtros salvos do CRM - Compatível com Meep"""
    return {
        "status": "success",
        "module": "crm",
        "endpoint": "saved_filters",
        "owner_id": owner_id,
        "filters": []
    }

# ============================================================================
# CAMPANHAS E PROMOÇÕES
# ============================================================================

@router.get("/campanhas")
async def listar_campanhas():
    """Lista todas as campanhas de marketing"""
    return {
        "status": "success",
        "module": "marketing",
        "endpoint": "campanhas",
        "campaigns": []
    }

@router.post("/campanhas")
async def criar_campanha():
    """Cria nova campanha de marketing"""
    return {
        "status": "success",
        "module": "marketing",
        "endpoint": "criar_campanha",
        "campaign_id": "camp_new_123"
    }

@router.get("/promocoes")
async def listar_promocoes():
    """Lista todas as promoções ativas"""
    return {
        "status": "success",
        "module": "marketing",
        "endpoint": "promocoes",
        "promotions": []
    }

@router.post("/promocoes")
async def criar_promocao():
    """Cria nova promoção"""
    return {
        "status": "success",
        "module": "marketing",
        "endpoint": "criar_promocao",
        "promotion_id": "promo_new_123"
    }

@router.get("/cupons")
async def listar_cupons():
    """Lista cupons de desconto"""
    return {
        "status": "success",
        "module": "marketing",
        "endpoint": "cupons",
        "coupons": []
    }

@router.post("/cupons")
async def criar_cupom():
    """Cria novo cupom de desconto"""
    return {
        "status": "success",
        "module": "marketing",
        "endpoint": "criar_cupom",
        "coupon_id": "coupon_new_123"
    }

@router.get("/descontos/lista-convidados")
async def listar_lista_convidados():
    """Lista de convidados para descontos especiais"""
    return {
        "status": "success",
        "module": "marketing",
        "endpoint": "lista_convidados",
        "guest_lists": []
    }

# ============================================================================
# ANALYTICS E MÉTRICAS
# ============================================================================

@router.get("/metricas/engajamento")
async def obter_metricas_engajamento():
    """Obtém métricas de engajamento das campanhas"""
    return {
        "status": "success",
        "module": "marketing",
        "endpoint": "metricas_engajamento",
        "metrics": {
            "total_campaigns": 25,
            "active_campaigns": 8,
            "avg_engagement_rate": 15.5,
            "total_customers_reached": 12500
        }
    }

@router.get("/analytics/conversao")
async def obter_taxa_conversao():
    """Obtém taxas de conversão das campanhas"""
    return {
        "status": "success",
        "module": "marketing",
        "endpoint": "conversao",
        "conversion_metrics": {
            "overall_rate": 12.3,
            "by_channel": {
                "email": 8.5,
                "sms": 15.2,
                "push": 18.7
            }
        }
    }

@router.get("/analytics/roi")
async def calcular_roi():
    """Calcula ROI das campanhas de marketing"""
    return {
        "status": "success",
        "module": "marketing",
        "endpoint": "roi",
        "roi_metrics": {
            "total_invested": 50000,
            "total_revenue": 125000,
            "roi_percentage": 150.0
        }
    }