# openapi_nip_full_20251022_021407.yaml
openapi: 3.1.0
info:
  title: MEEP Clone API
  version: 3.0.0
  description: 'Patch Plan — Primeira leva (Auth, Empresas, Usuários, Eventos, Clientes,
    PDV, Ingressos, Financeiro). Padrões aplicados: RESTful, JWT bearer, paginação,
    erros tipados, idempotência em operações financeiras, webhooks como POST, versionamento
    /api/v1.

    '
servers:
- url: https://api.seudominio.com/api/v1
security:
- bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ErrorResponse:
      description: Erro padrão
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: object
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
    SortParam:
      name: sort
      in: query
      schema:
        type: string
    QueryParam:
      name: q
      in: query
      schema:
        type: string
    DateFromParam:
      name: date_from
      in: query
      schema:
        type: string
        format: date-time
    DateToParam:
      name: date_to
      in: query
      schema:
        type: string
        format: date-time
...
  schemas:
...
    Venda:
      type: object
      properties:
        id:
          type: integer
        total:
          type: number
          format: float
        forma_pagamento:
          type: string
        itens:
          type: array
          items:
            type: object
            properties:
              produto_id:
                type: integer
              quantidade:
                type: number
              preco_unitario:
                type: number
                format: float
...
paths:
...
  /marketing/campanhas:
    get:
      tags:
      - Marketing
      summary: Listar marketing/campanhas
      parameters:
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/PageSizeParam'
      - $ref: '#/components/parameters/SortParam'
      - $ref: '#/components/parameters/QueryParam'
      responses:
        '200':
          description: OK
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags:
      - Marketing
      summary: Criar marketing/campanha
      responses:
        '201':
          description: Created
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /pdv/vendas:
    get:
      summary: Listar vendas
      parameters:
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/PageSizeParam'
      - $ref: '#/components/parameters/DateFromParam'
      - $ref: '#/components/parameters/DateToParam'
      responses:
        '200':
          description: OK
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      summary: Criar venda
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Venda'
      responses:
        '201':
          description: Created
        default:
          $ref: '#/components/responses/ErrorResponse'
...